requirements.txt
Django
requests
pandas
Pillow
icmplib
pycryptodome
plotly_express
kaleido
gunicorn
autopep8
isort
